Statusbar: Gracefully adapt 3Minit Battery hook for AOSP
3Minit Battery is a free or paid third party statusbar
battery icon solution. This commit builds upon the
initial SystemUI hook source release by Gary Harrington.

We are able to significantly optimize the original code
because we are not constrained by the "one-size-fits-all"
conditions the original code must meet. At the same time,
we fully maintain the integrity of the UX intended by
the original code.

Of note is the introduction of a controller class. This
class takes over settings and state management as well
as handles receiver events. The controller class also
monitors package changes and updates feature visibility
accordingly if a needed resource is added or removed.

Upon initialization, if no feature resources are
detected, all resources are freed.

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/KeyguardStatusBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java

Change-Id: If6d50f255ba46b392128b1c5f5baf8f27d00068a

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBarTransitions.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/StatusBarHeaderView.java